refactor: Move config params to core
