+++ web_ui/src/core/configurable-parameters/services/in-memory-api-model-config-parameters-service.ts
@@ -1,10 +1,10 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
-import { ConfigurableParametersTaskChain } from '../../../shared/components/configurable-parameters/configurable-parameters.interface';
 import { ProjectIdentifier } from '../../projects/core.interface';
 import { ConfigurableParametersReconfigureDTO } from '../dtos/configurable-parameters.interface';
 import { CreateApiModelConfigParametersService } from './api-model-config-parameters-service';
+import { ConfigurableParametersTaskChain } from './configurable-parameters.interface';
 import { mockedConfigParamData, mockedReadOnlyConfigTaskChainData } from './test-utils';
 
 export const createInMemoryApiModelConfigParametersService = (): CreateApiModelConfigParametersService => {
