+++ web_ui/src/pages/annotator/notification/auto-training-credits-modal/auto-training-credits-modal.test.tsx
@@ -3,12 +3,12 @@
 
 import { fireEvent, screen, waitFor, waitForElementToBeRemoved } from '@testing-library/react';
 
+import { BooleanGroupParams } from '../../../../core/configurable-parameters/services/configurable-parameters.interface';
 import { JobState } from '../../../../core/jobs/jobs.const';
 import { createInMemoryJobsService } from '../../../../core/jobs/services/in-memory-jobs-service';
 import { FUX_NOTIFICATION_KEYS, FUX_SETTINGS_KEYS } from '../../../../core/user-settings/dtos/user-settings.interface';
 import { UserGlobalSettings, UseSettings } from '../../../../core/user-settings/services/user-settings.interface';
 import { INITIAL_GLOBAL_SETTINGS } from '../../../../core/user-settings/utils';
-import { BooleanGroupParams } from '../../../../shared/components/configurable-parameters/configurable-parameters.interface';
 import { useAutoTrainingTasksConfig } from '../../../../shared/components/header/active-learning-configuration/use-tasks-auto-training-config.hook';
 import { getFuxSetting } from '../../../../shared/components/tutorials/utils';
 import { getMockedJob, getMockedJobCount } from '../../../../test-utils/mocked-items-factory/mocked-jobs';
