+++ web_ui/src/pages/project-details/components/project-models/legacy-train-model-dialog/use-training-state-value/use-training-state-value.hook.tsx
@@ -5,14 +5,14 @@ import { useCallback, useMemo, useState } from 'react';
 
 import { isNumber } from 'lodash-es';
 
+import { ConfigurableParametersTaskChain } from '../../../../../../core/configurable-parameters/services/configurable-parameters.interface';
+import { updateSelectedParameter } from '../../../../../../core/configurable-parameters/utils';
 import { useFeatureFlags } from '../../../../../../core/feature-flags/hooks/use-feature-flags.hook';
 import { TrainingBodyDTO } from '../../../../../../core/models/dtos/train-model.interface';
 import { useModels } from '../../../../../../core/models/hooks/use-models.hook';
 import { Task } from '../../../../../../core/projects/task.interface';
 import { useTasksWithSupportedAlgorithms } from '../../../../../../core/supported-algorithms/hooks/use-tasks-with-supported-algorithms';
 import { AnimationDirections } from '../../../../../../shared/animation-parameters/animation-parameters';
-import { ConfigurableParametersTaskChain } from '../../../../../../shared/components/configurable-parameters/configurable-parameters.interface';
-import { updateSelectedParameter } from '../../../../../../shared/components/configurable-parameters/utils';
 import { isNotCropTask } from '../../../../../../shared/utils';
 import { useTotalCreditPrice } from '../../../../hooks/use-credits-to-consume.hook';
 import { useProject } from '../../../../providers/project-provider/project-provider.component';
