+++ web_ui/src/pages/project-details/components/project-models/legacy-train-model-dialog/utils.ts
@@ -1,8 +1,8 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
+import { ConfigurableParametersTaskChain } from '../../../../../core/configurable-parameters/services/configurable-parameters.interface';
 import { TrainingBodyDTO } from '../../../../../core/models/dtos/train-model.interface';
-import { ConfigurableParametersTaskChain } from '../../../../../shared/components/configurable-parameters/configurable-parameters.interface';
 import { getTrainingConfigParametersDTO } from '../utils';
 
 interface TrainingBody {
