+++ web_ui/src/pages/project-details/components/project-models/train-model-dialog/advanced-settings/evaluation/evaluation.component.tsx
@@ -9,7 +9,7 @@ import { noop } from 'lodash-es';
 import {
     ConfigurableParametersParams,
     NumberGroupParams,
-} from '../../../../../../../shared/components/configurable-parameters/configurable-parameters.interface';
+} from '../../../../../../../core/configurable-parameters/services/configurable-parameters.interface';
 import { Accordion } from '../ui/accordion/accordion.component';
 import { NumberParameter } from '../ui/number-parameter.component';
 import { ResetButton } from '../ui/reset-button.component';
