+++ web_ui/src/pages/project-details/components/project-models/train-model-dialog/advanced-settings/training/training.component.tsx
@@ -5,7 +5,7 @@ import { FC } from 'react';
 
 import { View } from '@geti/ui';
 
-import { ConfigurableParametersTaskChain } from '../../../../../../../shared/components/configurable-parameters/configurable-parameters.interface';
+import { ConfigurableParametersTaskChain } from '../../../../../../../core/configurable-parameters/services/configurable-parameters.interface';
 import { FineTuneParameters } from './fine-tune-parameters.component';
 import { LearningParameters } from './learning-parameters.component';
