+++ web_ui/src/pages/project-details/components/project-models/train-model-dialog/advanced-settings/ui/parameters.component.tsx
@@ -6,7 +6,7 @@ import { FC, ReactNode } from 'react';
 import { Grid, minmax, Text, View } from '@geti/ui';
 import { isFunction } from 'lodash-es';
 
-import { ConfigurableParametersParams } from '../../../../../../../shared/components/configurable-parameters/configurable-parameters.interface';
+import { ConfigurableParametersParams } from '../../../../../../../core/configurable-parameters/services/configurable-parameters.interface';
 import { BooleanParameter } from './boolean-parameter.component';
 import { NumberParameter } from './number-parameter.component';
 import { ResetButton } from './reset-button.component';
