+++ web_ui/src/shared/components/configurable-parameters/cp-group-item/cp-group-item.component.tsx
@@ -6,8 +6,11 @@ import { useState } from 'react';
 import { ActionButton, Flex, Text, View } from '@geti/ui';
 import { ChevronRightSmallLight, ChevronUpLight } from '@geti/ui/icons';
 
+import {
+    ConfigParameterItemProp,
+    ConfigurableParametersGroups,
+} from '../../../../core/configurable-parameters/services/configurable-parameters.interface';
 import { idMatchingFormat } from '../../../../test-utils/id-utils';
-import { ConfigParameterItemProp, ConfigurableParametersGroups } from '../configurable-parameters.interface';
 import { CPParamsList } from '../cp-list/cp-list.component';
 
 import classes from './cp-group-item.module.scss';
