+++ web_ui/src/shared/components/configurable-parameters/cp-groups-list/cp-groups-list.component.tsx
@@ -4,8 +4,11 @@
 import { View } from '@geti/ui';
 import { AnimatePresence, motion } from 'framer-motion';
 
+import {
+    ConfigParameterItemProp,
+    ConfigurableParametersGroups,
+} from '../../../../core/configurable-parameters/services/configurable-parameters.interface';
 import { ANIMATION_PARAMETERS } from '../../../animation-parameters/animation-parameters';
-import { ConfigParameterItemProp, ConfigurableParametersGroups } from '../configurable-parameters.interface';
 import { CPGroupItem } from '../cp-group-item/cp-group-item.component';
 
 interface CPGroupsListProps extends ConfigParameterItemProp {
