+++ web_ui/src/shared/components/configurable-parameters/cp-item/cp-editable-item/cp-editable-item.interface.ts
@@ -1,7 +1,10 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
-import { ConfigurableParametersParams, ConfigurableParametersSingle } from '../../configurable-parameters.interface';
+import {
+    ConfigurableParametersParams,
+    ConfigurableParametersSingle,
+} from '../../../../../core/configurable-parameters/services/configurable-parameters.interface';
 
 export interface ResetButtonHandler extends Pick<ConfigurableParametersSingle, 'updateParameter'> {
     id: string;
