+++ web_ui/src/shared/components/configurable-parameters/cp-item/cp-editable-item/cp-number/cp-number.component.tsx
@@ -6,7 +6,7 @@ import { useMemo } from 'react';
 import { Flex, NumberField, Slider, Text, useMediaQuery, useNumberFormatter } from '@geti/ui';
 import { isLargeSizeQuery } from '@geti/ui/theme';
 
-import { NumberGroupParams } from '../../../configurable-parameters.interface';
+import { NumberGroupParams } from '../../../../../../core/configurable-parameters/services/configurable-parameters.interface';
 import { ResetButtonHandler } from '../cp-editable-item.interface';
 
 import classes from './cp-number.module.scss';
