+++ web_ui/src/shared/components/configurable-parameters/cp-item/utils.ts
@@ -1,7 +1,7 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
-import { ConfigurableParametersParams } from '../configurable-parameters.interface';
+import { ConfigurableParametersParams } from '../../../../core/configurable-parameters/services/configurable-parameters.interface';
 
 export const getStaticContent = (parameter: ConfigurableParametersParams): string | number => {
     let content: string | number = '';
