+++ web_ui/src/shared/components/configurable-parameters/selectable-many-tasks/selectable-many-tasks.interface.ts
@@ -1,7 +1,7 @@
 // Copyright (C) 2022-2025 Intel Corporation
 // LIMITED EDGE SOFTWARE DISTRIBUTION LICENSE
 
-import { ConfigurableParametersMany } from '../configurable-parameters.interface';
+import { ConfigurableParametersMany } from '../../../../core/configurable-parameters/services/configurable-parameters.interface';
 
 export interface SelectableManyTasksProps extends Omit<ConfigurableParametersMany, 'type' | 'configParametersData'> {
     configurableParameters: ConfigurableParametersMany['configParametersData'];
