+++ web_ui/src/shared/components/configurable-parameters/selectable-many-tasks/selected-tab-content/selectable-tab-content.component.tsx
@@ -3,7 +3,7 @@
 
 import { View } from '@geti/ui';
 
-import { ConfigurableParametersMany } from '../../configurable-parameters.interface';
+import { ConfigurableParametersMany } from '../../../../../core/configurable-parameters/services/configurable-parameters.interface';
 import { CPDescription } from '../../cp-description/cp-description.component';
 import { CPGroupsList } from '../../cp-groups-list/cp-groups-list.component';
 import { CPParamsList } from '../../cp-list/cp-list.component';
