+++ web_ui/src/shared/components/header/active-learning-configuration/training-settings/required-annotations-slider.component.tsx
@@ -6,12 +6,12 @@ import { FC, useState } from 'react';
 import { Flex, Slider, Text, View } from '@geti/ui';
 
 import { useReconfigAutoTraining } from '../../../../../core/configurable-parameters/hooks/use-reconfig-auto-training.hook';
-import { findRequiredImagesAutoTrainingConfig } from '../../../../../core/configurable-parameters/utils';
-import { Task } from '../../../../../core/projects/task.interface';
 import {
     ConfigurableParametersTaskChain,
     NumberGroupParams,
-} from '../../../configurable-parameters/configurable-parameters.interface';
+} from '../../../../../core/configurable-parameters/services/configurable-parameters.interface';
+import { findRequiredImagesAutoTrainingConfig } from '../../../../../core/configurable-parameters/utils';
+import { Task } from '../../../../../core/projects/task.interface';
 
 interface RequiredAnnotationsSliderProps {
     task: Task;
